move traits into submodules
